# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "拆镜学 API",
    "description": "拆镜学原型后端：上传参考视频 -> 拆解（镜头边界/关键帧/台词/景别/画面描述）-> 选择学习目标 -> 生成可完成的学习任务 -> 提交作品 -> 结构差异与证据化反馈。所有对外部模型的调用都发生在服务端。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project_id}/uploads": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "申请上传地址（短时有效）",
        "operationId": "create_upload_api_v1_projects__project_id__uploads_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadTicketOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}/content": {
      "put": {
        "tags": [
          "assets"
        ],
        "summary": "上传原始文件（签名地址或 Bearer 令牌）",
        "operationId": "upload_content_api_v1_assets__asset_id__content_put",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}/complete": {
      "post": {
        "tags": [
          "assets"
        ],
        "summary": "确认上传并触发服务端媒体验证",
        "operationId": "complete_upload_api_v1_assets__asset_id__complete_post",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediaValidationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{asset_id}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "素材详情（含短期播放地址）",
        "operationId": "get_asset_api_v1_assets__asset_id__get",
        "parameters": [
          {
            "name": "asset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Asset Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/projects/{project_id}/assets": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "项目素材列表",
        "operationId": "list_assets_api_v1_projects__project_id__assets_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetOut"
                  },
                  "title": "Response List Assets Api V1 Projects  Project Id  Assets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/media/{token}": {
      "get": {
        "tags": [
          "assets"
        ],
        "summary": "短期媒体访问（默认 5 分钟有效）",
        "operationId": "get_media_api_v1_media__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "duration_ms": {
            "type": "integer",
            "title": "Duration Ms"
          },
          "width": {
            "type": "integer",
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "title": "Height"
          },
          "fps": {
            "type": "number",
            "title": "Fps"
          },
          "has_audio": {
            "type": "boolean",
            "title": "Has Audio"
          },
          "video_codec": {
            "type": "string",
            "title": "Video Codec"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "is_vfr": {
            "type": "boolean",
            "title": "Is Vfr"
          },
          "rejection_code": {
            "type": "string",
            "title": "Rejection Code",
            "default": ""
          },
          "rejection_reason": {
            "type": "string",
            "title": "Rejection Reason",
            "default": ""
          },
          "playback_url": {
            "type": "string",
            "title": "Playback Url",
            "default": ""
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "status",
          "duration_ms",
          "width",
          "height",
          "fps",
          "has_audio",
          "video_codec",
          "media_type",
          "size_bytes",
          "sha256",
          "is_vfr",
          "created_at"
        ],
        "title": "AssetOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MediaValidationOut": {
        "properties": {
          "asset": {
            "$ref": "#/components/schemas/AssetOut"
          },
          "validated": {
            "type": "boolean",
            "title": "Validated"
          },
          "checks": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Checks"
          }
        },
        "type": "object",
        "required": [
          "asset",
          "validated",
          "checks"
        ],
        "title": "MediaValidationOut"
      },
...
      "UploadRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Filename"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Size Bytes"
          },
          "media_type": {
            "type": "string",
            "maxLength": 64,
            "title": "Media Type",
            "default": "video/mp4"
          },
          "sha256": {
            "type": "string",
            "maxLength": 64,
            "title": "Sha256",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "filename",
          "size_bytes"
        ],
        "title": "UploadRequest"
      },
      "UploadTicketOut": {
        "properties": {
          "asset_id": {
            "type": "string",
            "title": "Asset Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "object_key": {
            "type": "string",
            "title": "Object Key"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "upload_method": {
            "type": "string",
            "title": "Upload Method",
            "default": "PUT"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          },
          "max_bytes": {
            "type": "integer",
            "title": "Max Bytes"
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "asset_id",
          "project_id",
          "object_key",
          "upload_url",
          "expires_in",
          "max_bytes"
        ],
        "title": "UploadTicketOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}